+++ src/main.rs
@@ -22,9 +22,4 @@ fn main() {
     }
 
     log::debug!("？？？「幾重にも辛酸を舐め、七難八苦を超え、艱難辛苦の果て、満願成就に至る——」");
-
-    use game::Game;
-
-    let game = Game::new();
-    game.execute();
 }
