+++ src/renderer/vulkano/mod.rs
@@ -1,3 +1,6 @@
+pub mod surface;
+pub mod instance;
+
 use crate::renderer::*;
 
 pub struct VulkanoBackend;
