+++ src/renderer/vulkano/surface.rs
@@ -1,0 +1,216 @@
+use vulkano::command_buffer::DynamicState;
+use vulkano::device::{Device, DeviceExtensions, Queue};
+use vulkano::framebuffer::{Framebuffer, FramebufferAbstract, RenderPassAbstract};
+use vulkano::image::swapchain::SwapchainImage;
+use vulkano::instance::PhysicalDevice;
+use vulkano::pipeline::viewport::Viewport;
+use vulkano::swapchain::{FullscreenExclusive, PresentMode, Surface, SurfaceTransform, Swapchain};
+
+use winit::event::{Event, WindowEvent};
+use winit::event_loop::{ControlFlow, EventLoop};
+use winit::window::{Window, WindowBuilder};
+
+use std::sync::Arc;
+
+use crate::constants;
+use super::instance;
+
+pub struct VulkanSurface<'a> {
+    pub physical: PhysicalDevice<'a>,
+    pub device: Arc<Device>,
+    pub surface: Arc<Surface<Window>>,
+    pub swapchain: Arc<Swapchain<Window>>,
+    pub images: Vec<Arc<SwapchainImage<Window>>>,
+    pub graphical_queue: Arc<Queue>,
+    pub transfer_queue: Arc<Queue>,
+}
+
+impl VulkanSurface<'static> {
+    pub fn new(event_loop: &EventLoop<()>) -> Self {
+        /*
+         * Vulkan-based program should follow these instructions to ininitalize:
+         *
+         * - Create an instance
+         * - Obtain a physical device
+         * - Create a Vulkan surface from Window
+         *   - This requires the creation of a winit Window.
+         * - Create a device
+         */
+        
+        let physical = Self::create_physical();
+        let surface = Self::create_window(event_loop);
+        let (device, graphical_queue, transfer_queue) = Self::create_device(physical, &surface);
+
+        let caps = surface.capabilities(physical).unwrap();
+        use vulkano::format::Format;
+
+        log::debug!("supported formats: {:?}", caps.supported_formats);
+
+        let (f, cs) = caps
+            .supported_formats
+            .iter()
+            .copied()
+            .find(|(f, _)| {
+                *f == Format::B8G8R8A8Srgb
+                    || *f == Format::B8G8R8Srgb
+                    || *f == Format::R8G8B8A8Srgb
+                    || *f == Format::R8G8B8Srgb
+            })
+            .unwrap();
+
+        let (swapchain, images) = Swapchain::new(
+            device.clone(),
+            surface.clone(),
+            caps.min_image_count,
+            f,
+            surface.window().inner_size().into(),
+            1,
+            caps.supported_usage_flags,
+            &graphical_queue,
+            SurfaceTransform::Identity,
+            caps.supported_composite_alpha.iter().next().unwrap(),
+            PresentMode::Fifo,
+            FullscreenExclusive::Default,
+            true,
+            cs,
+        )
+        .expect("failed to create a swapchain");
+
+        VulkanSurface {
+            physical,
+            device,
+            surface,
+            swapchain,
+            images,
+            graphical_queue,
+            transfer_queue,
+        }
+    }
+}
+
+// -- Initialization
+impl<'a> VulkanSurface<'a> {
+    fn create_physical() -> PhysicalDevice<'static> {
+        let instance = instance::get_instance();
+
+        // Obtain a physical device.
+        // Note that a PhysicalDevice is bound to the reference of the instance,
+        // hence the instance should be alive while `physical` is alive.
+        // Instance has 'static lifetime parameter, so no problem here.
+        //
+        // TODO: let users to choose physical devices
+        let physical = PhysicalDevice::enumerate(instance)
+            .next()
+            .expect("no physical device available");
+
+        log::debug!("device: {}, type: {:?}", physical.name(), physical.ty());
+
+        physical
+    }
+
+    fn create_window(event_loop: &EventLoop<()>) -> Arc<Surface<Window>> {
+        use winit::dpi::LogicalSize;
+
+        let window = WindowBuilder::new()
+            .with_title(constants::GAME_ENGINE_FULL_NAME)
+            .with_inner_size(LogicalSize {
+                width: constants::GAME_WINDOW_WIDTH,
+                height: constants::GAME_WINDOW_HEIGHT,
+            })
+            // .with_resizable(false)
+            .build(event_loop)
+            .expect("failed to build Window");
+
+        log::debug!(
+            "created Window; size: {}, {}",
+            constants::GAME_WINDOW_WIDTH,
+            constants::GAME_WINDOW_HEIGHT
+        );
+
+        let surface = vulkano_win::create_vk_surface(window, instance::get_instance().clone())
+            .expect("failed to build Vulkan surface");
+
+        log::debug!("created Vulkan surface");
+
+        surface
+    }
+
+    fn create_device<T>(
+        physical: PhysicalDevice,
+        surface: &Surface<T>,
+    ) -> (Arc<Device>, Arc<Queue>, Arc<Queue>) {
+        let gr_queue_family = physical
+            .queue_families()
+            .find(|&q| q.supports_graphics() && surface.is_supported(q).unwrap_or(false));
+
+        let tr_queue_family = physical.queue_families().find(|&q| {
+            (q.supports_graphics() || q.supports_compute()) // VK_QUEUE_TRANSFER_BIT
+                && gr_queue_family != Some(q) // no overlap
+        });
+
+        let extensions = DeviceExtensions {
+            khr_swapchain: true, // swapchain is required
+            ..DeviceExtensions::none()
+        };
+
+        let (d, mut q) = Device::new(
+            physical,
+            physical.supported_features(),
+            &extensions,
+            vec![(gr_queue_family, 1.0), (tr_queue_family, 0.5)]
+                .into_iter()
+                .filter_map(|(v, a)| Some((v?, a))),
+        )
+        .expect("failed to create device");
+
+        // graphics queue and transfer queue
+        let gq = q.next().unwrap();
+        let tq = q.next().unwrap_or_else(|| gq.clone());
+
+        log::debug!("created device and queue");
+
+        (d, gq, tq)
+    }
+}
+
+// -- Run-loop execution & event-handling
+impl VulkanSurface<'static> {
+    /// Executes an event loop.
+    ///
+    /// It takes the ownership of a Game instance, and won't return until
+    /// the program is closed.
+    pub fn execute(mut self) {
+    }
+
+    pub fn perform_redraw(&mut self) {
+        // TODO:
+    }
+}
+
+fn window_size_dependent_setup(
+    images: &[Arc<SwapchainImage<Window>>],
+    render_pass: Arc<dyn RenderPassAbstract + Send + Sync>,
+    dynamic_state: &mut DynamicState,
+) -> Vec<Arc<dyn FramebufferAbstract + Send + Sync>> {
+    let dimensions = images[0].dimensions();
+
+    let viewport = Viewport {
+        origin: [0.0, 0.0],
+        dimensions: [dimensions[0] as f32, dimensions[1] as f32],
+        depth_range: 0.0..1.0,
+    };
+    dynamic_state.viewports = Some(vec![viewport]);
+
+    images
+        .iter()
+        .map(|image| {
+            Arc::new(
+                Framebuffer::start(render_pass.clone())
+                    .add(image.clone())
+                    .unwrap()
+                    .build()
+                    .unwrap(),
+            ) as Arc<dyn FramebufferAbstract + Send + Sync>
+        })
+        .collect::<Vec<_>>()
+}
